Revised error check for aperture size < pixel size
Previously, a stop was triggered if the minimum aperture size was smaller than the pixel size (e.g. due to inclination). In the current version of the code (which allows the minimum aperture size to increase if needed), this no longer makes sense. A warning has been added, together with an increase of the minimum aperture size, and the stop is restricted to cases where less than 2 aperture sizes exceed the inclination-corrected pixel diagonal. The use of diagonal rather than size implies the addition of a factor of sqrt(2).